Rename `medm` directory to `flowam`
