+++ python/tk_multi_loader/build_asset_dialog.py
@@ -15,8 +15,8 @@ from sgtk.platform.qt import QtGui
 from tank_vendor.flow_integration_sdk.exceptions import FlowError
 from tank_vendor.flow_integration_sdk.objects import FlowProject
 
-from .medm.template_queries import get_template_pipeline_steps, get_templates
-from .medm.create import CreateMode, get_template_source_path
+from .flowam.template_queries import get_template_pipeline_steps, get_templates
+from .flowam.create import CreateMode, get_template_source_path
 from .ui.build_asset_dialog import Ui_BuildAssetDialog
 
 # Toolkit logger
